pull out config related functions and middleware
